+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2025-03-04
+
+### Added
+ - Placeholder constant for ffl license number. 
+ - Search capability on FFL license management settings page, new Search tab.
+ - Added new Modify tab for future work. This will have ability to edit FFL license info in db.
+ 
 ## [1.1.1] - 2025-02-27
 
 ### Removed